Move debug gui to core
